04_Code/synesthesia2/requirements.txt
# SYNESTHESIA 2.0 Requirements

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
Pillow>=9.0.0

# Audio processing
librosa>=0.9.0
soundfile>=0.10.0

# Deep learning (Phase 2)
torch>=1.10.0
torchvision>=0.11.0
timm>=0.5.0

# GPU acceleration (optional)
# cupy-cuda11x  # Uncomment for NVIDIA GPU support

# Video processing
ffmpeg-python>=0.2.0

# 3D visualization (optional, for advanced rendering)
pyvista>=0.34.0
# open3d>=0.14.0

# Web interface (Phase 4 - optional)
# gradio>=3.0.0
# streamlit>=1.10.0
